reverting datamanagement changes due to them being impossible
